Prepare scala-json for new releases for Scala 2.10 and 2.11
Problem

There is no scala-json release for Scala 2.11. The primary obstacle is the fact
that there is no specs release for Scala 2.11, and the project uses specs for
its tests. There are also minor issues with the project's documentation, build,
licensing, etc.

Solution

I've migrated the tests from specs to ScalaTest, updated the documentation and
licenses, removed 2.9 from the SBT build, and added configuration for a Travis
CI build. The test migration is as literal as possible, and does not attempt to
address any issues with the specs tests apart from the fact that they were
written in specs.

Result

The project builds (with SBT) and tests pass for both 2.10 and 2.11. There are
no changes to the project code (apart from the tests) except for one addition to
a header and one stylistic change (de-relativizing an import).

RB_ID=534101